feat: add role checking option for any WordPress user role
Handling of middleware functions that are invoked before the route callable has changed. From now on, if a variable is returned from a middleware function and that variable is not empty, PHP will die with the result. This functionality was added to avoid having to call die directly inside middleware functions, which made it difficult if not impossible to test these functions.